requirements.txt
# Core Streamlit and Web Framework
streamlit>=1.28.0
requests>=2.31.0

# AI and Language Models
groq>=0.3.0

# Document Processing
PyPDF2>=3.0.1
python-docx>=0.8.11

# Web Scraping and Automation
selenium>=4.15.0
playwright>=1.40.0
beautifulsoup4>=4.12.0
aiohttp==3.9.0
nest-asyncio>=1.6.0

# Jira Integration
jira>=3.5.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
plotly==5.18.0

# Environment Management
python-dotenv>=1.0.0

# Image Processing and Computer Vision
pillow>=10.0.0
opencv-python>=4.8.0
scikit-image>=0.21.0
matplotlib>=3.7.0

# Additional Utilities
webdriver-manager>=4.0.0
urllib3>=2.0.0
certifi>=2023.7.22

# Optional: For enhanced image comparison
scipy>=1.11.0
imageio>=2.31.0

# Optional: to export HTML reports to PDF (requires wkhtmltopdf installed on system)
pdfkit>=1.0.0

# Development and Testing (Optional)
pytest>=7.4.0
# pytest-streamlit>=1.2.0
black>=23.7.0
flake8>=6.0.0